+++ server/src/main/java/org/allaymc/server/entity/component/EntityDamageComponentImpl.java
@@ -249,7 +249,7 @@ public class EntityDamageComponentImpl implements EntityDamageComponent {
         }
 
         var blockStateStandingOn = thisEntity.getBlockStateStandingOn().blockState();
-        float rawDamage = (event.getFallDistance() - 3) - baseComponent.getEffectLevel(EffectTypes.JUMP_BOOST);
+        double rawDamage = (event.getFallDistance() - 3) - baseComponent.getEffectLevel(EffectTypes.JUMP_BOOST);
         var damage = Math.round(rawDamage * (1 - blockStateStandingOn.getBehavior().getFallDamageReductionFactor()));
         if (damage > 0) {
             attack(DamageContainer.fall(damage));
